refactor(hooks): extract business logic from screens to custom hooks
Extract API interactions, state management, and handlers from screen
components to enforce the thin-screen pattern:
- Move SMS logic to useSMSData, useSMSActions, and useMessageLinks
- Move update logic to useUpdateCheck and useUpdateDownload
- Move login logic to useLogin
- Move settings logic to useSystemSettings, useLanSettings, and useMobileNetwork
- Reduce overall screen file sizes in src/app by ~47%
